New method for calculating hamming does NOT require compression now, so unit64 -> int.

+++ src/cpp/platformConfig.h
@@ -21,7 +21,7 @@
     using LoopIndexType = size_t;
     using randomGenerator = pcg32;
 
-    inline int countBits(uint64_t x) {
+    inline int countBits(int x) {
         return __builtin_popcountll(x);
     }
